+++ src/pages/LeaderboardPage/components/LeaderboardHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from './LeaderboardHeader'
